corregido

+++ sql/create-table-users.sql
@@ -4,7 +4,7 @@ create Table users(
     user_name VARCHAR(99) NOT NULL,
     user_email VARCHAR(99) NOT NULL UNIQUE,
     user_password VARCHAR(99) NOT NULL,
-    user_state int(1) DEFAULT 1,
+    user_state integer(1) DEFAULT 1,
     user_time_created DATETIME DEFAULT CURRENT_TIMESTAMP,
     user_time_updated DATETIME DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
